+++ iOS/swiftui-oidc/PingExample/PingExample/Views/ConfigurationView.swift
@@ -12,7 +12,7 @@ import SwiftUI
 import FRAuth
 
 struct ConfigurationView: View {
-    @Binding var viewmodel: ConfigurationViewModel
+    @Binding var configurationViewModel: ConfigurationViewModel
     @State private var scopes: String = ""
     @State private var environments = ["AIC", "PingOne"]
     @State private var selectedEnvironment = "AIC"
@@ -23,25 +23,25 @@ struct ConfigurationView: View {
             Section(header: Text("OAuth 2.0 details")) {
                 Section {
                     Text("Client Id:")
-                    TextField("Client Id", text: $viewmodel.clientId)
+                    TextField("Client Id", text: $configurationViewModel.clientId)
                         .disableAutocorrection(true)
                         .autocapitalization(.none)
                 }
                 Section {
                     Text("Redirect URI:")
-                    TextField("Redirect URI", text: $viewmodel.redirectUri)
+                    TextField("Redirect URI", text: $configurationViewModel.redirectUri)
                         .disableAutocorrection(true)
                         .autocapitalization(.none)
                 }
                 Section {
                     Text("SignOut URI:")
-                    TextField("SignOut URI", text: $viewmodel.signOutUri.toUnwrapped(defaultValue: ""))
+                    TextField("SignOut URI", text: $configurationViewModel.signOutUri.toUnwrapped(defaultValue: ""))
                         .disableAutocorrection(true)
                         .autocapitalization(.none)
                 }
                 Section {
                     Text("Discovery Endpoint:")
-                    TextField("Discovery Endpoint", text: $viewmodel.discoveryEndpoint)
+                    TextField("Discovery Endpoint", text: $configurationViewModel.discoveryEndpoint)
                         .disableAutocorrection(true)
                         .autocapitalization(.none)
                 }
@@ -59,13 +59,13 @@ struct ConfigurationView: View {
                             Text(env)
                         }
                     }.onChange(of: selectedEnvironment) { env in
-                        viewmodel.environment = env
+                        configurationViewModel.environment = env
                     }
                     .pickerStyle(.menu)
                 }
                 Section {
                     Text("Cookie Name (Optional):")
-                    TextField("Cookie Name:", text: $viewmodel.cookieName.toUnwrapped(defaultValue: ""))
+                    TextField("Cookie Name:", text: $configurationViewModel.cookieName.toUnwrapped(defaultValue: ""))
                         .disableAutocorrection(true)
                         .autocapitalization(.none)
                 }
@@ -76,25 +76,25 @@ struct ConfigurationView: View {
                         Text(type.rawValue).tag(type.rawValue)
                     }
                 }.onChange(of: selectedBrowserType) { type in
-                    viewmodel.browserSeletorType = type
+                    configurationViewModel.browserSeletorType = type
                 }
                 .pickerStyle(.menu)
             }
             Section(header: Text("Browser Configuration")) {
                 Button(action: {
                     Task {
-                        viewmodel.scopes = scopes.components(separatedBy: " ")
-                        viewmodel.saveConfiguration()
-                        await viewmodel.startSDK()
+                        configurationViewModel.scopes = scopes.components(separatedBy: " ")
+                        configurationViewModel.saveConfiguration()
+                        await configurationViewModel.startSDK()
                     }
                 }) {
                     Text("Save Configuration")
                 }
                 Button(action: {
                     Task {
-                        let defaultConfiguration = viewmodel.resetConfiguration()
-                        viewmodel = defaultConfiguration
-                        scopes = $viewmodel.scopes.wrappedValue.joined(separator: " ")
+                        let defaultConfiguration = configurationViewModel.resetConfiguration()
+                        configurationViewModel = defaultConfiguration
+                        scopes = $configurationViewModel.scopes.wrappedValue.joined(separator: " ")
                     }
                 }) {
                     Text("Reset Configuration")
@@ -103,15 +103,15 @@ struct ConfigurationView: View {
         }
         .navigationTitle("Edit Configuration")
         .onAppear{
-            scopes = $viewmodel.scopes.wrappedValue.joined(separator: " ")
-            selectedBrowserType = viewmodel.browserSeletorType
-            selectedEnvironment = viewmodel.environment
+            scopes = $configurationViewModel.scopes.wrappedValue.joined(separator: " ")
+            selectedBrowserType = configurationViewModel.browserSeletorType
+            selectedEnvironment = configurationViewModel.environment
         }
         .onDisappear{
-            viewmodel.scopes = scopes.components(separatedBy: " ")
-            viewmodel.saveConfiguration()
+            configurationViewModel.scopes = scopes.components(separatedBy: " ")
+            configurationViewModel.saveConfiguration()
             Task {
-                await viewmodel.startSDK()
+                await configurationViewModel.startSDK()
             }
         }
     }
